refactor(vald): extract evm.go into multiple files
