+++ vald/evm/gateway_tx_confirmation.go
@@ -1,0 +1,115 @@
+package evm
+
+import (
+	"bytes"
+	"context"
+
+	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
+	"github.com/ethereum/go-ethereum/common"
+	geth "github.com/ethereum/go-ethereum/core/types"
+
+	"github.com/axelarnetwork/axelar-core/x/evm/types"
+	nexus "github.com/axelarnetwork/axelar-core/x/nexus/exported"
+	voteTypes "github.com/axelarnetwork/axelar-core/x/vote/types"
+)
+
+// ProcessGatewayTxConfirmation votes on the correctness of an EVM chain gateway's transactions
+func (mgr Mgr) ProcessGatewayTxConfirmation(event *types.ConfirmGatewayTxStarted) error {
+	if !mgr.isParticipantOf(event.Participants) {
+		mgr.logger("pollID", event.PollID).Debug("ignoring gateway tx confirmation poll: not a participant")
+		return nil
+	}
+
+	txReceipt, err := mgr.GetTxReceiptIfFinalized(event.Chain, common.Hash(event.TxID), event.ConfirmationHeight)
+	if err != nil {
+		return err
+	}
+	if txReceipt == nil {
+		mgr.logger().Infof("broadcasting empty vote for poll %s", event.PollID.String())
+		_, err := mgr.broadcaster.Broadcast(context.TODO(), voteTypes.NewVoteRequest(mgr.proxy, event.PollID, types.NewVoteEvents(event.Chain)))
+
+		return err
+	}
+
+	events := mgr.processGatewayTxLogs(event.Chain, event.GatewayAddress, txReceipt.Logs)
+
+	mgr.logger().Infof("broadcasting vote %v for poll %s", events, event.PollID.String())
+	_, err = mgr.broadcaster.Broadcast(context.TODO(), voteTypes.NewVoteRequest(mgr.proxy, event.PollID, types.NewVoteEvents(event.Chain, events...)))
+
+	return err
+}
+
+// extract receipt processing from ProcessGatewayTxConfirmation, so that it can be used in ProcessGatewayTxsConfirmation
+func (mgr Mgr) processGatewayTxLogs(chain nexus.ChainName, gatewayAddress types.Address, logs []*geth.Log) []types.Event {
+	var events []types.Event
+	for i, txlog := range logs {
+		if !bytes.Equal(gatewayAddress.Bytes(), txlog.Address.Bytes()) {
+			continue
+		}
+
+		switch txlog.Topics[0] {
+		case ContractCallSig:
+			gatewayEvent, err := DecodeEventContractCall(txlog)
+			if err != nil {
+				mgr.logger().Debug(sdkerrors.Wrap(err, "decode event ContractCall failed").Error())
+				continue
+			}
+
+			if err := gatewayEvent.ValidateBasic(); err != nil {
+				mgr.logger().Debug(sdkerrors.Wrap(err, "invalid event ContractCall").Error())
+				continue
+			}
+
+			events = append(events, types.Event{
+				Chain: chain,
+				TxID:  types.Hash(txlog.TxHash),
+				Index: uint64(i),
+				Event: &types.Event_ContractCall{
+					ContractCall: &gatewayEvent,
+				},
+			})
+		case ContractCallWithTokenSig:
+			gatewayEvent, err := DecodeEventContractCallWithToken(txlog)
+			if err != nil {
+				mgr.logger().Debug(sdkerrors.Wrap(err, "decode event ContractCallWithToken failed").Error())
+				continue
+			}
+
+			if err := gatewayEvent.ValidateBasic(); err != nil {
+				mgr.logger().Debug(sdkerrors.Wrap(err, "invalid event ContractCallWithToken").Error())
+				continue
+			}
+
+			events = append(events, types.Event{
+				Chain: chain,
+				TxID:  types.Hash(txlog.TxHash),
+				Index: uint64(i),
+				Event: &types.Event_ContractCallWithToken{
+					ContractCallWithToken: &gatewayEvent,
+				},
+			})
+		case TokenSentSig:
+			gatewayEvent, err := DecodeEventTokenSent(txlog)
+			if err != nil {
+				mgr.logger().Debug(sdkerrors.Wrap(err, "decode event TokenSent failed").Error())
+			}
+
+			if err := gatewayEvent.ValidateBasic(); err != nil {
+				mgr.logger().Debug(sdkerrors.Wrap(err, "invalid event TokenSent").Error())
+				continue
+			}
+
+			events = append(events, types.Event{
+				Chain: chain,
+				TxID:  types.Hash(txlog.TxHash),
+				Index: uint64(i),
+				Event: &types.Event_TokenSent{
+					TokenSent: &gatewayEvent,
+				},
+			})
+		default:
+		}
+	}
+
+	return events
+}
